Added RedditCacheService and FileService
The RedditCache service is responsible to manage the local cache for the json retrieved from the web, it has all the methods that should be inside a cache class, store, retrieve and invalidate. The FileService is a wrapper class for the Default FileManager and it is used by the Reddit Cache for store, retrieve and delete cache files.